Split local DB review tests
